Add ability to display alternative (or no) content instead of redirecting
Ideal for preventing the display of a widget for specific users. If the
`ContentPermissions` part is configured with no redirect URL it's up to
the view template for the content type to control what's displayed to on
the page.

Added a liquid filter for checking permission of content item against
the user.

(#5)